Filter feature bug fix

+++ videogamestore-java-code/src/main/java/org/yearup/controllers/OrdersController.java
@@ -44,10 +44,14 @@ public class OrdersController {
         if (shoppingCart.getItems().isEmpty()) {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR);
         }
+        if (profileDao.getByUserId(userId).getAddress().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
         Order order = orderDao.createOrder(userId, profile, shoppingCart);
 
-        for (Map.Entry<Integer, ShoppingCartItem> shoppingCartItem: shoppingCart.getItems().entrySet()) {
-            orderLineItemDao.create(shoppingCartItem.getKey(), order.getOrderId(), shoppingCartItem.getValue());
+        for (ShoppingCartItem item : shoppingCart.getItems().values()) {
+            int productId = item.getProduct().getProductId();
+            orderLineItemDao.create(productId, order.getOrderId(), item);
         }
         shoppingCartDao.deleteShoppingCart(userId);
         return order;
